+++ src/main/java/uk/gov/hmcts/reform/fis/services/S2sAuthService.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.cosapi.services;
+package uk.gov.hmcts.reform.fis.services;
 
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.stereotype.Service;
