+++ src/test/java/uk/gov/hmcts/reform/fis/services/S2sAuthServiceTest.java
@@ -1,4 +1,4 @@
-package uk.gov.hmcts.reform.cosapi.services;
+package uk.gov.hmcts.reform.fis.services;
 
 import org.junit.jupiter.api.Assertions;
 import org.junit.jupiter.api.Test;
